refactor: player ID from PeerId, start_game assigns deploy zones and colors
- NetState gains local_player_id: u16, set after deriving from PeerId
- lobby: remove host_color_index, derive player ID via derive_local_player_id()
- context: start_game takes explicit local/peer player IDs, assigns deploy zones and colors from PlayerState; MatchProgress::new takes &[u16]
- Fix cascading local_player_id: u8 → u16 in main, phase_ui, rendering, chat, ui

+++ src/chat.rs
@@ -7,7 +7,7 @@ use crate::team;
 pub struct ChatMessage {
     pub name: String,
     pub text: String,
-    pub player_id: u8,
+    pub player_id: u16,
     pub lifetime: f32,
 }
 
@@ -46,7 +46,7 @@ impl ChatState {
         phase: &GamePhase,
         net: &mut Option<net::NetState>,
         player_name: &str,
-        local_id: u8,
+        local_id: u16,
     ) {
         let chat_allowed = matches!(
             phase,
